refactor: 테스트 셋업 코드 분리
